change: Split TaskCommand and TasksCommand into two modules
