+++ src/gmp/gmp.ts
@@ -46,7 +46,8 @@ import {PortListCommand, PortListsCommand} from 'gmp/commands/portlists';
 import {ReportCommand, ReportsCommand} from 'gmp/commands/reports';
 import RoleCommand from 'gmp/commands/role';
 import RolesCommand from 'gmp/commands/roles';
-import {TaskCommand, TasksCommand} from 'gmp/commands/tasks';
+import TaskCommand from 'gmp/commands/task';
+import TasksCommand from 'gmp/commands/tasks';
 import TrashCanCommand from 'gmp/commands/trashcan';
 import UserCommand from 'gmp/commands/user';
 import UsersCommand from 'gmp/commands/users';
